requirements.txt
pycvesearch
python-nmap
pyfiglet
paramiko
